requirements.txt
# Azure Vector Databases Demo - Python Dependencies

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Database Connectivity
pyodbc>=4.0.35
sqlalchemy>=2.0.0

# Azure Cosmos DB Integration
azure-cosmos>=4.5.0

# MongoDB Integration (for Cosmos DB MongoDB vCore)
pymongo>=4.6.0
motor>=3.3.0  # Async MongoDB driver (optional)

# Azure OpenAI Integration
openai>=1.12.0

# Text Processing and Tokenization
tiktoken>=0.5.0

# Machine Learning (for TF-IDF fallback option)
scikit-learn>=1.3.0

# Jupyter Notebook Support
jupyter>=1.0.0
ipykernel>=6.20.0

# Development and Utilities
python-dotenv>=1.0.0  # For environment variable management
requests>=2.28.0      # HTTP requests
urllib3>=1.26.0       # URL handling

# Data Visualization (optional)
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.10.0

# Performance and Monitoring
psutil>=5.9.0         # System monitoring
tqdm>=4.64.0          # Progress bars

# Security and Configuration
cryptography>=40.0.0  # Secure connections
python-jose>=3.3.0    # JWT handling (if needed)

# Testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0

# Code Quality (optional)
black>=23.0.0         # Code formatting
flake8>=6.0.0         # Linting
mypy>=1.0.0           # Type checking
